+++ src/load_inputs/load_minimum_capacity_requirement.jl
@@ -1,7 +1,7 @@
 @doc raw"""
     load_minimum_capacity_requirement!(path::AbstractString, inputs::Dict, setup::Dict)
 
-Read input parameters related to mimimum capacity requirement constraints (e.g. technology specific deployment mandates)
+Read input parameters related to minimum capacity requirement constraints (e.g. technology specific deployment mandates)
 """
 function load_minimum_capacity_requirement!(path::AbstractString, inputs::Dict, setup::Dict)
     filename = "Minimum_capacity_requirement.csv"