+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Include comprehensive list of all settings flags (that are currently in GenX) in the docs (#721)
 - Allow `load_vre_stor_variability!` to load only wind or solar resources as 
 part of the VRE_STOR module (#728)
+- Fix typo in HydrogenMinimumProduction setting key (#746)
 
 ### Added
 - Add objective scaler for addressing problem ill-conditioning (#667)